+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## 0.3.0 - 2015-04-16
+
+  + fix bug with node template not showing os icon
+  + when clicking a table in a node view - add select query to editor
+  + Truncate node metadata in node listing.
+  + Editor comments, new lines and pop last valid for current search.
+  + Way better error messages for query syntax issues.
+  + Add a few new default queries:
+    * Third-party kernel extensions (OS X)
+    * Startup items (OS X)
+    * Interface information
+    * Shell history
+    * All users with groups
+  + minor bug fixes
+
+  - socket verbose output for errors returned from handlers
+
+## 0.2.4 - 2015-04-16
+
+  + better debug output for node request/time of query
+  + daemon support for server
+  + minor bug fixes
+
 ## 0.2.3 - 2015-04-13
 
   + os icon for node button